+++ extensions/sn_extra_game_options/change_log.md
@@ -4,3 +4,5 @@ Change Log
 * 1.0
   - Split off from simple_menu_api
   - Added option to hide help text popups.
+* 1.1
+  - Added option to focus a newly opened map on player